VRT reclassify: Split functionality into separate class
