+++ frmts/vrt/CMakeLists.txt
@@ -19,6 +19,8 @@ add_gdal_driver(
           vrtprocesseddataset.cpp
           vrtprocesseddatasetfunctions.cpp
           vrtmultidim.cpp
+          vrtreclassifier.h
+          vrtreclassifier.cpp
           STRONG_CXX_WFLAGS)
 gdal_standard_includes(gdal_vrt)
 target_include_directories(gdal_vrt PRIVATE ${GDAL_RASTER_FORMAT_SOURCE_DIR}/raw
